+++ rs/src/tags.rs
@@ -104,8 +104,7 @@ pub struct DefineDynamicText {
   pub color: Option<StraightSRgba8>,
   #[serde(skip_serializing_if = "Option::is_none")]
   pub max_length: Option<usize>,
-  #[serde(skip_serializing_if = "Option::is_none")]
-  pub align: Option<TextAlignment>,
+  pub align: TextAlignment,
   pub margin_left: u16,
   pub margin_right: u16,
   pub indent: u16,